backend/requirements.txt

flask
flask-cors
flask-sqlalchemy
flask-jwt-extended
flask-mail
python-dotenv
werkzeug
flask-dance
oauthlib
